fix: bubbleSort

+++ src/easySort.cpp
@@ -51,7 +51,7 @@ vector<int> bubbleSort(vector<int> arr){
 		return arr;
 	}
 
-	for(int i = arr.size() - 1; i > 0; --i){
+	for(int i = (int)arr.size() - 1; i > 0; --i){
 		for(int j = 0; j < i; ++j){
 			if(arr[j] > arr[i]){
 				arr = swap(arr, i, j);
